feat: 新增打点员的ID权限验证，修正部分注释

+++ genshin-map-api/genshin-map-api-core/genshin-map-api-core-core/src/main/java/site/yuanshen/genshin/core/controller/PunctuateController.java
@@ -37,18 +37,20 @@ public class PunctuateController {
         );
     }
 
-    @Operation(summary = "分页查询自己提交的未通过的打点信息（不包含额外字段）",
-            description = "分页查询自己提交的未通过的打点信息（不包含额外字段）（打点员的API）")
-    @PostMapping("/get/page_single/{authorId}")
-    public R<PageListVo<MarkerPunctuateVo>> listSelfSinglePunctuatePage(@RequestBody PageSearchVo pageSearchVo, @PathVariable("authorId") Long authorId) {
+    @Operation(summary = "分页查询自己提交的未通过的打点信息",
+            description = "分页查询自己提交的未通过的打点信息（打点员的API）")
+    @PostMapping("/get/page/{authorId}")
+    public R<PageListVo<MarkerPunctuateVo>> listSelfPunctuatePage(@RequestBody PageSearchVo pageSearchVo, @PathVariable("authorId") Long authorId, @RequestHeader("userId") Long userId) {
+        if (!userId.equals(authorId)) throw new RuntimeException("无权限查看其他人未提交的打点信息");
         return RUtils.create(
                 punctuateService.listSelfPunctuatePage(new PageSearchDto(pageSearchVo), authorId)
         );
     }
 
-    @Operation(summary = "提交暂存点位（不含额外字段）", description = "成功则返回打点提交ID")
-    @PutMapping("/single")
-    public R<Long> addSinglePunctuate(@RequestBody MarkerPunctuateVo markerPunctuateVo) {
+    @Operation(summary = "提交暂存点位", description = "成功则返回打点提交ID")
+    @PutMapping("/")
+    public R<Long> addSinglePunctuate(@RequestBody MarkerPunctuateVo markerPunctuateVo, @RequestHeader("userId") Long userId) {
+        if (!userId.equals(markerPunctuateVo.getAuthor())) throw new RuntimeException("无权限为其他人提交打点信息");
         return RUtils.create(
                 punctuateService.addSinglePunctuate(new MarkerPunctuateDto(markerPunctuateVo))
         );
@@ -56,15 +58,17 @@ public class PunctuateController {
 
     @Operation(summary = "将暂存点位提交审核", description = "将暂存点位提交审核")
     @PutMapping("/push/{authorId}")
-    public R<Boolean> pushPunctuate(@PathVariable("authorId") Long authorId) {
+    public R<Boolean> pushPunctuate(@PathVariable("authorId") Long authorId, @RequestHeader("userId") Long userId) {
+        if (!userId.equals(authorId)) throw new RuntimeException("无权限将其他用户的暂存点位提交审核");
         return RUtils.create(
                 punctuateService.pushPunctuate(authorId)
         );
     }
 
-    @Operation(summary = "修改自身未提交的暂存点位（不包括额外字段）", description = "根据点位ID修改点位")
-    @PostMapping("/single")
-    public R<Boolean> updateSelfSinglePunctuate(@RequestBody MarkerPunctuateVo markerPunctuateVo) {
+    @Operation(summary = "修改自身未提交的暂存点位", description = "根据点位ID修改点位")
+    @PostMapping("/")
+    public R<Boolean> updateSelfSinglePunctuate(@RequestBody MarkerPunctuateVo markerPunctuateVo, @RequestHeader("userId") Long userId) {
+        if (!userId.equals(markerPunctuateVo.getAuthor())) throw new RuntimeException("无权限修改其他人未提交的打点信息");
         return RUtils.create(
                 punctuateService.updateSelfSinglePunctuate(new MarkerPunctuateDto(markerPunctuateVo))
         );
@@ -72,7 +76,8 @@ public class PunctuateController {
 
     @Operation(summary = "删除自己未通过的提交点位", description = "根据提交ID列表来删除提交点位，会对打点员ID进行校验")
     @DeleteMapping("/delete/{authorId}/{punctuateId}")
-    public R<Boolean> deleteSelfPunctuate(@PathVariable("punctuateId") Long punctuateId, @PathVariable("authorId") Long authorId) {
+    public R<Boolean> deleteSelfPunctuate(@PathVariable("punctuateId") Long punctuateId, @PathVariable("authorId") Long authorId, @RequestHeader("userId") Long userId) {
+        if (!userId.equals(authorId)) throw new RuntimeException("无权限删除其他用户的暂存点位");
         return RUtils.create(
                 punctuateService.deleteSelfPunctuate(punctuateId, authorId)
         );
